Move utilities into UserHooks
This moves some utilities for dealing with UserHooks into that module to
avoid circular imports.
