+++ Cabal/Distribution/Simple/UserHooks.hs
@@ -23,22 +23,27 @@
 -- various phases because it would involve changing the types of the
 -- corresponding hook. At some point it will have to be replaced.
 
-module Distribution.Simple.UserHooks (
-        UserHooks(..), Args,
-        emptyUserHooks,
-  ) where
+module Distribution.Simple.UserHooks
+       ( UserHooks(..), Args, emptyUserHooks
+       , hookedAction, hookedActionWithArgs
+       , sanityCheckHookedBuildInfo
+       ) where
 
 import Distribution.PackageDescription
-         (PackageDescription, GenericPackageDescription,
-          HookedBuildInfo, emptyHookedBuildInfo)
-import Distribution.Simple.Program    (Program)
+    ( PackageDescription(..), GenericPackageDescription
+    , Executable(..), HookedBuildInfo, emptyHookedBuildInfo
+    , updatePackageDescription )
 import Distribution.Simple.Command    (noExtraFlags)
+import Distribution.Simple.LocalBuildInfo (LocalBuildInfo(..))
 import Distribution.Simple.PreProcess (PPSuffixHandler)
+import Distribution.Simple.Program    (Program)
 import Distribution.Simple.Setup
          (ConfigFlags, BuildFlags, ReplFlags, CleanFlags, CopyFlags,
           InstallFlags, SDistFlags, RegisterFlags, HscolourFlags,
           HaddockFlags, TestFlags, BenchmarkFlags)
-import Distribution.Simple.LocalBuildInfo (LocalBuildInfo)
+import Distribution.Simple.Utils (die)
+
+import Data.List (nub, (\\))
 
 type Args = [String]
 
@@ -209,3 +214,58 @@ emptyUserHooks
     where rn  args _ = noExtraFlags args >> return emptyHookedBuildInfo
           rn' _    _ = return emptyHookedBuildInfo
           ru _ _ _ _ = return ()
+
+hookedAction :: (UserHooks -> Args -> flags -> IO HookedBuildInfo)
+                -- ^ pre-hook
+             -> (UserHooks -> PackageDescription -> LocalBuildInfo
+                 -> UserHooks -> flags -> IO ())
+                -- ^ command hook
+             -> (UserHooks -> Args -> flags -> PackageDescription
+                 -> LocalBuildInfo -> IO ())
+                -- ^ post-hook
+             -> IO LocalBuildInfo
+             -> UserHooks -> flags -> Args -> IO ()
+hookedAction pre_hook cmd_hook =
+    hookedActionWithArgs pre_hook $ \h _ pd lbi uh flags ->
+                                      cmd_hook h pd lbi uh flags
+
+hookedActionWithArgs :: (UserHooks -> Args -> flags -> IO HookedBuildInfo)
+                        -- ^ pre-hook
+                     -> (UserHooks -> Args -> PackageDescription
+                         -> LocalBuildInfo -> UserHooks -> flags -> IO ())
+                        -- ^ command hook, with arguments
+                     -> (UserHooks -> Args -> flags -> PackageDescription
+                         -> LocalBuildInfo -> IO ())
+                        -- ^ post-hook
+                     -> IO LocalBuildInfo
+                     -> UserHooks -> flags -> Args -> IO ()
+hookedActionWithArgs
+  pre_hook cmd_hook post_hook get_build_config
+  hooks flags args = do
+    pbi <- pre_hook hooks args flags
+    localbuildinfo <- get_build_config
+    let pkg_descr0 = localPkgDescr localbuildinfo
+    --pkg_descr0 <- get_pkg_descr (get_verbose flags)
+    sanityCheckHookedBuildInfo pkg_descr0 pbi
+    let pkg_descr = updatePackageDescription pbi pkg_descr0
+    -- TODO: should we write the modified package descr back to the
+    -- localbuildinfo?
+    cmd_hook hooks args pkg_descr localbuildinfo hooks flags
+    post_hook hooks args flags pkg_descr localbuildinfo
+
+sanityCheckHookedBuildInfo :: PackageDescription -> HookedBuildInfo -> IO ()
+sanityCheckHookedBuildInfo PackageDescription { library = Nothing } (Just _,_)
+    = die $ "The buildinfo contains info for a library, "
+         ++ "but the package does not have a library."
+
+sanityCheckHookedBuildInfo pkg_descr (_, hookExes)
+    | not (null nonExistant)
+    = die $ "The buildinfo contains info for an executable called '"
+         ++ head nonExistant ++ "' but the package does not have a "
+         ++ "executable with that name."
+  where
+    pkgExeNames  = nub (map exeName (executables pkg_descr))
+    hookExeNames = nub (map fst hookExes)
+    nonExistant  = hookExeNames \\ pkgExeNames
+
+sanityCheckHookedBuildInfo _ _ = return ()
